fix(cli): carry --checksum-seed as a signed int (#7561)
Upstream declares the seed as `int checksum_seed` (options.c:151), parses it
with `POPT_ARG_INT` (options.c:861), and forwards it to the remote with `"%d"`
(options.c:3047). popt bounds `POPT_ARG_INT` to `INT_MIN..=INT_MAX` and rejects
anything outside with `POPT_ERROR_OVERFLOW` (popt/popt.c poptSaveArg).

The `--server` / remote-shell path carried the seed as `u32`, which broke both
directions against upstream 3.5.0:

  - `oc-rsync --checksum-seed=4294967295` was accepted and emitted
    `--checksum-seed=4294967295`, which an upstream server refuses with
    "number too large or too small" (exit 1) before any bytes move.
  - `oc-rsync --server --checksum-seed=-1` exited 1 with "must be
    0..4294967295", so an upstream client run with a negative seed could not
    talk to an oc-rsync server. Upstream accepts the same argv.
  - `oc-rsync --checksum-seed=-1` was refused outright, so oc-rsync could not
    express a value upstream accepts.

The daemon's own client-arg parser already parsed the seed as `i32` and
documented why; it just had to cast back to `u32` to store it. This was an
internal inconsistency as well as an upstream divergence.

Retype the user-specified seed as `i32` from the CLI parse through
`ClientConfig`, the remote-shell and daemon argv emitters, `ServerConfig`,
`ProtocolSetupConfig` and `write_seed`. The wire bytes are unchanged: the seed
was already reinterpreted as `i32` before `write_int`, so this fixes the
rendered text and the two parse ends, not the protocol.

Sign extension into the xxhash seed now matches upstream too: checksum.c:342-350
passes the signed `int` into `XXH64` / `XXH3_*_withSeed`, where C sign-extends
to 64 bits.

+++ crates/daemon/src/daemon/sections/module_access/client_args/long_form_args.rs
@@ -482,13 +482,13 @@ fn apply_long_form_args(
                 // on the wire. Dropping the value made `--checksum-seed` a no-op
                 // over a daemon while it worked over a remote shell.
                 //
-                // Parsed as `i32` and reinterpreted, because that is the width
-                // upstream prints from and the width `write_seed` puts back on
-                // the wire; a `u32`-only parse would reject the negative values
-                // upstream's `%d` can emit.
+                // Parsed as `i32`, because that is the width upstream prints
+                // from and the width `write_seed` puts back on the wire; a
+                // `u32`-only parse would reject the negative values upstream's
+                // `%d` can emit.
                 } else if let Some(seed) = arg.strip_prefix("--checksum-seed=") {
                     if let Ok(value) = seed.trim().parse::<i32>() {
-                        config.checksum_seed = Some(value as u32);
+                        config.checksum_seed = Some(value);
                     }
                 } else if arg == "--from0" {
                     // upstream: options.c:940 - --from0 sets NUL-delimited mode
